feat/refactor: configurações globais da API e ajustes de domínio em Pedido

+++ src/main/java/com/devaguiar/erp/dtos/responses/ProdutoResponseDTO.java
@@ -2,7 +2,9 @@ package com.devaguiar.erp.dtos.responses;
 
 import com.devaguiar.erp.entities.Produto;
 
-public record ProdutoResponseDTO(Long id, String nome, Double preco, Integer quantidade) {
+import java.math.BigDecimal;
+
+public record ProdutoResponseDTO(Long id, String nome, BigDecimal preco, Integer quantidade) {
 
     public ProdutoResponseDTO(Produto produto) {
         this(produto.getId(), produto.getNome(), produto.getPreco(), produto.getQuantidade());
